Working on send message button.
