+++ Concertify/package.json
@@ -33,11 +33,10 @@
     "expo-system-ui": "~4.0.7",
     "expo-web-browser": "~14.0.2",
     "firebase": "^11.0.1",
+    "fs": "^0.0.1-security",
     "fuse": "^0.12.1",
-=========
     "fuse.js": "^7.0.0",
     "prop-types": "^15.8.1",
->>>>>>>>> Temporary merge branch 2
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-native": "0.76.6",
